+++ src/mesoscopy/io.py
@@ -18,14 +18,17 @@
 #  IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR
 #  IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 #  SOFTWARE.
+import csv
 import typing
+import xmltodict
+from collections import OrderedDict
 
 import h5py
 import zarr
 
+import numpy as np
 import numpy.typing as npt
 from dask import array as da
-
 from pynwb import NWBHDF5IO, NWBFile
 
 
@@ -119,7 +122,78 @@ def store_interim(
             dtype=array.dtype,
             chunks=(chunks, array.shape[1], array.shape[2]),
         )
-        return array.store(z_interim, return_stored=True, compute=compute)
+        return da.store(array, z_interim, return_stored=True, compute=compute)
 
     zarr.save(interim_path, array)
     return zarr.load(interim_path)
+
+
+def read_points(path: str) -> dict[str, tuple[float, float]]:
+    """Read a landmark points file.
+
+    Args:
+        path (str): Path to the points file.
+
+    Returns:
+        dict[str, tuple[float, float]]: Dictionary with the landmark names as keys and their x-y coordinates
+    """
+    if path.endswith(".xml") or path.endswith(".points"):
+        return _read_fiji_points(path)
+    if path.endswith(".csv"):
+        return _read_csv_points(path)
+    raise ValueError("Unsupported file format.")
+
+
+def _read_fiji_points(path: str) -> dict[str, tuple[float, float]]:
+    """Read a FIJI landmark points file.
+
+    Args:
+        path (str): Path to the points file.
+
+    Returns:
+        dict[str, tuple[float, float]]: Dictionary with the landmark names as keys and their x-y coordinates
+    """
+    with open(path, "r") as fp:
+        points = xmltodict.parse(fp.read())
+        points = OrderedDict(
+            {
+                point["@name"]: (float(point["@y"]), float(point["@x"]))
+                for point in points["namedpointset"]["pointworld"]
+            }
+        )
+
+    return points
+
+
+def _read_csv_points(path: str) -> dict[str, tuple[float, float]]:
+    """Read a CSV file with landmark points.
+
+    Args:
+        path (str): Path to the CSV file.
+
+    Returns:
+        dict[str, tuple[float, float]]: Dictionary with the landmark names as keys and their x-y coordinates
+    """
+    with open(path, "r") as fp:
+        csv_reader = csv.DictReader(fp)
+        points = OrderedDict(
+            {row["landmark"]: (float(row["y"]), float(row["x"])) for row in csv_reader}
+        )
+
+    return points
+
+
+def write_points(path: str, points: dict[str, tuple[float, float]]) -> None:
+    """Write a dictionary of landmark points to a CSV file.
+
+    Args:
+        path (str): Path to output CSV file.
+        points (dict[str, tuple[float, float]]): Dictionary with the landmark names as keys and their x-y coordinates
+    """
+    if not path.endswith(".csv"):
+        path += ".csv"
+    with open(path, "w") as fp:
+        csv_writer = csv.DictWriter(fp, fieldnames=["landmark", "y", "x"])
+        csv_writer.writeheader()
+        for landmark, (y, x) in points.items():
+            csv_writer.writerow({"landmark": landmark, "y": y, "x": x})